Extract PreviewFrame component
Most of the `Preview` component was concerned with catching runtime
errors and dealing with cross-browser-compatible frame rendering (Safari
and IE need a different approach than Chrome and Firefox). Extract that
code out into a `PreviewFrame` that deals with the mechanics of frame
building and communication but is ignorant of the business logic of the
frame.
